requirements.txt
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
bleach==5.0.1
build==0.8.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
Deprecated==1.2.13
docutils==0.18.1
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
keyring==23.6.0
multidict==6.0.2
packaging==21.3
pep517==0.12.0
pkginfo==1.8.3
pycparser==2.21
PyGithub==1.55
Pygments==2.12.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-gitlab==3.6.0
readme-renderer==35.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
six==1.16.0
smmap==5.0.0
tomli==2.0.1
tqdm==4.64.0
twine==4.0.1
typing_extensions==4.3.0
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
